refactor: complete codebase revamp
- Implemented new folder structure for better maintainability
- Modularized components (PatternCard, PatternGrid, etc.)
- Added custom hooks for theme, favorites, and copy-to-clipboard
- Centralized app constants and data files
- Preserved all original functionality
