+++ src/app/not-found.tsx
@@ -1,0 +1,27 @@
+import Link from "next/link";
+import { Button } from "@/components/ui/button";
+import { Home } from "lucide-react";
+
+export default function NotFound() {
+  return (
+    <div className="min-h-screen flex items-center justify-center">
+      <div className="text-center">
+        <h1 className="text-6xl font-bold text-gray-900 dark:text-gray-100 mb-4">
+          404
+        </h1>
+        <h2 className="text-2xl font-semibold text-gray-700 dark:text-gray-300 mb-4">
+          Page Not Found
+        </h2>
+        <p className="text-gray-600 dark:text-gray-400 mb-8">
+          The page you&apos;re looking for doesn&apos;t exist.
+        </p>
+        <Button asChild>
+          <Link href="/" className="flex items-center gap-2">
+            <Home className="h-4 w-4" />
+            Go Home
+          </Link>
+        </Button>
+      </div>
+    </div>
+  );
+}
