+++ src/app/page.tsx
@@ -1,43 +1,30 @@
 "use client";
 
-import Hero from "./components/hero";
-import Navbar from "./components/navbar";
-import Footer from "./components/footer";
-import PatternShowcase from "./components/pattern-showcase";
 import { useState, useEffect } from "react";
-import { ThemeProvider } from "./components/theme-provider";
-import { gridPatterns } from "./utils/patterns";
 import { Toaster } from "sonner";
-import SupportDropdown from "./components/SupportDropdownProps ";
-import ReturnToPreview from "./components/ReturnToPreview";
+import { ThemeProvider } from "@/components/providers/theme-provider";
+import { gridPatterns } from "@/data/patterns";
+import { useTheme } from "@/hooks/useTheme";
+import Navbar from "@/components/layout/navbar";
+import Footer from "@/components/layout/footer";
+import Hero from "@/components/home/hero";
+import PatternShowcase from "@/components/patterns/pattern-showcase";
+import SupportDropdown from "@/components/home/support-dropdown";
+import ReturnToPreview from "@/components/home/return-to-preview";
 
 export default function Home() {
   const [activePattern, setActivePattern] = useState<string | null>(null);
-  const [theme, setTheme] = useState<"light" | "dark">("light");
+  const { theme, updateThemeFromPattern } = useTheme();
+
+  // Update theme based on pattern background color
+  useEffect(() => {
+    updateThemeFromPattern(activePattern, gridPatterns);
+  }, [activePattern, updateThemeFromPattern]);
 
   // Find the active pattern object
   const activePatternObj = activePattern
     ? gridPatterns.find((p) => p.id === activePattern)
     : null;
-
-  // Update theme based on pattern background color
-  useEffect(() => {
-    if (activePatternObj) {
-      // Check if pattern ID starts with "dark-" or contains specific dark colors
-      const background = activePatternObj.style.background || "";
-      const isDark =
-        activePatternObj.id.startsWith("dark-") ||
-        (typeof background === "string" &&
-          (background.includes("#0") ||
-            background.includes("#1") ||
-            background.includes("rgba(0,") ||
-            background.includes("rgba(1,")));
-
-      setTheme(isDark ? "dark" : "light");
-    } else {
-      setTheme("light");
-    }
-  }, [activePattern, activePatternObj]);
 
   return (
     <>
@@ -49,7 +36,7 @@ export default function Home() {
           )}
           <div className="relative z-10">
             <Navbar theme={theme} />
-            <SupportDropdown theme={theme}/>
+            <SupportDropdown theme={theme} />
             <Hero
               activePattern={activePattern}
               setActivePattern={setActivePattern}
@@ -62,7 +49,7 @@ export default function Home() {
             />
             <Footer theme={theme} />
           </div>
-          <ReturnToPreview theme={theme}  />
+          <ReturnToPreview theme={theme} />
         </div>
         <Toaster />
       </ThemeProvider>
