+++ src/data/categories.ts
@@ -1,0 +1,5 @@
+import { PATTERN_CATEGORIES } from "@/lib/constants";
+
+export const categories = PATTERN_CATEGORIES;
+
+export type CategoryId = (typeof PATTERN_CATEGORIES)[number]["id"];
